+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v0.32 (2025-11-xx)
 
+New features:
+
+- Added Claude Opus 4.5 support via the Anthropic AI backend.
+- Added Kimi-k2, Minimax-m2, GLM-4.6, and Gemini3-Pro-Preview via the Ollama AI backend.  Updated GLM and Gemini3 3 via the
+  Z.ai and Google backends to differentiate the two ways to access the same things.
+
 ## v0.31 (2025-11-25)
 
 New features: